Moved normalized store name to a more proper module
