+++ packages/cli-kit/src/environment/fqdn.ts
@@ -1,5 +1,5 @@
-import {fqdn as spinFqdn} from './spin.js'
-import {serviceEnvironment} from './service.js'
+import {fqdn, fqdn as spinFqdn} from './spin.js'
+import {isSpinEnvironment, serviceEnvironment} from './service.js'
 import {Abort} from '../error.js'
 
 export const CouldntObtainPartnersSpinFQDNError = new Abort(
@@ -68,4 +68,17 @@ export async function shopify(options: {storeFqdn?: string} = {}): Promise<strin
         throw NotProvidedStoreFQDNError
       }
   }
+}
+
+/**
+ * Given a store, returns a valid store fqdn removing protocol and adding the proper domain in case is missing
+ * @param store - Original store name provided by the user
+ * @returns a valid store fqdn
+ */
+export function normalizeStoreName(store: string) {
+  const storeFqdn = store.replace(/^https?:\/\//, '').replace(/\/$/, '')
+  const addDomain = async (storeFqdn: string) =>
+    isSpinEnvironment() ? `${storeFqdn}.shopify.${await fqdn()}` : `${storeFqdn}.myshopify.com`
+  const containDomain = (storeFqdn: string) => storeFqdn.includes('.myshopify.com') || storeFqdn.includes('spin.dev')
+  return containDomain(storeFqdn) ? storeFqdn : addDomain(storeFqdn)
 }
