+++ packages/cli-kit/src/session.ts
@@ -2,7 +2,7 @@ import {applicationId} from './session/identity.js'
 import {Abort, Bug} from './error.js'
 import {validateSession} from './session/validate.js'
 import {allDefaultScopes, apiScopes} from './session/scopes.js'
-import {identity as identityFqdn, partners as partnersFqdn} from './environment/fqdn.js'
+import {identity as identityFqdn, normalizeStoreName, partners as partnersFqdn} from './environment/fqdn.js'
 import {open} from './system.js'
 import {
   exchangeAccessForApplicationTokens,
@@ -21,7 +21,6 @@ import {authorize} from './session/authorize.js'
 import {IdentityToken, Session} from './session/schema.js'
 import * as secureStore from './session/store.js'
 import constants from './constants.js'
-import {normalizeStoreName} from './string.js'
 import * as output from './output.js'
 import {partners} from './api.js'
 import {RequestClientError} from './api/common.js'
