+++ packages/cli-kit/src/string.test.ts
@@ -1,39 +1,5 @@
-import {hashString, linesToColumns, normalizeStoreName, tryParseInt} from './string.js'
+import {hashString, linesToColumns, tryParseInt} from './string.js'
 import {describe, expect, it} from 'vitest'
-
-describe('normalizeStore', () => {
-  it('parses store name with http', async () => {
-    // When
-    const got = await normalizeStoreName('http://example.myshopify.com')
-
-    // Then
-    expect(got).toEqual('example.myshopify.com')
-  })
-
-  it('parses store name with https', async () => {
-    // When
-    const got = await normalizeStoreName('https://example.myshopify.com')
-
-    // Then
-    expect(got).toEqual('example.myshopify.com')
-  })
-
-  it('parses store name with https when spin URL', async () => {
-    // When
-    const got = await normalizeStoreName('https://devstore001.shopify.partners-6xat.test.us.spin.dev')
-
-    // Then
-    expect(got).toEqual('devstore001.shopify.partners-6xat.test.us.spin.dev')
-  })
-
-  it('parses store name without domain', async () => {
-    // When
-    const got = await normalizeStoreName('example')
-
-    // Then
-    expect(got).toEqual('example.myshopify.com')
-  })
-})
 
 describe('tryParseInt', () => {
   it('converts a string to an int', () => {
